Moved jump points out of the params list.
The reason is that params are meant to be things that either ARE values,
or CONTAIN values. A jump point is closer to a label or symbol than a
parameter.

It's possible down the line that jump points will simply be converted
into values behind the scenes. If that becmes the case then maybe I'll
move them back. For now I think it makes more sense to separate them.